pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "swissgeol-assets-dataextraction"
version = "0.1.0"
description = "Python project for geological document analysis."
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "boto3>=1.39.14,<2.0.0",
    "numpy>=2.0.0,<3.0.0",
    "matplotlib>=3.10.3,<4.0.0",
    "opencv-python>=4.12.0.88,<5.0.0.0",
    "pandas>=2.3.1,<3.0.0",
    "pillow>=11.3.0,<12.0.0",
    "pymupdf>=1.26.3,<2.0.0",
    "layoutparser>=0.3.4,<0.4.0",
    "regex>=2024.11.6,<2025.0.0",
    "scipy==1.14.1",
    "tabulate>=0.9.0,<0.10.0",
    "tqdm>=4.67.1,<5.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "shapely>=2.1.1,<3.0.0",
    "scikit-image>=0.25.2,<0.26.0",
    "scikit-learn>=1.7.1,<2.0.0",
    "joblib>=1.5.1,<2.0.0",
    "xgboost>=3.0.2,<4.0.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "fasttext-predict>=0.9.2.4,<0.10.0.0",
    "torch>=2.7.1,<3.0.0",
    "transformers>=4.54.0,<5.0.0"
]

[project.optional-dependencies]
test = [
    "pytest>=8.4.1,<9.0.0",
]
lint = [
    "pre-commit>=4.2.0,<5.0.0",
    "ruff"
]
experiment-tracking = [
    "mlflow>=3.1.4,<4.0.0",
    "pygit2>=1.18.1,<2.0.0"
]
deep-learning = [
    "datasets>=4.0.0,<5.0.0",
    "accelerate>=1.9.0,<2.0.0"
]

[tool.ruff]
line-length = 119

[tool.ruff.lint]
select = ["D", "E", "F", "UP", "B", "SIM", "I"]
ignore = ["D102", "D105", "D107", "D100"]

[tool.ruff.lint.pydocstyle]
convention = "google"
